Changed datatype of two properties

+++ models/line.rb
@@ -2,8 +2,8 @@ class Line
   include DataMapper::Resource
 
   property :id, Serial
-  property :latitude, Integer
-  property :longitude, Integer
+  property :latitude, Float
+  property :longitude, Float
   property :sequence_pos, Integer
   property :free_move, Boolean
